Add files via upload
Mountains are hard to climb and mountaineers require passion and perseverance to ascend. I decided to draw a mountain scene to reflect the passion and perseverance of the early leaders (the ballers) of this project. Much like climbing a mountain, the work of a baller requires a conscious effort to move out of comfort zones and risk failing. Their commitment, focus and vision gets them to the summit.